Fix change-of-sign warnings caught by ICC 17
  error #68: integer conversion resulted in a change of sign

+++ src/plugins/platforms/eglfs/deviceintegration/eglfs_kms_egldevice/qeglfskmsegldevicescreen.cpp
@@ -88,7 +88,7 @@ void QEglFSKmsEglDeviceScreen::waitForFlip()
 
         qCDebug(qLcEglfsKmsDebug, "Setting mode");
         int ret = drmModeSetCrtc(device()->fd(), output().crtc_id,
-                                 -1, 0, 0,
+                                 uint32_t(-1), 0, 0,
                                  &output().connector_id, 1,
                                  &output().modes[output().mode]);
         if (ret)
